Some fixes

+++ gaiku-3d/examples/common/exporter.rs
@@ -2,7 +2,7 @@ use obj_exporter::{Geometry, ObjSet, Object, Primitive, Shape, Vertex};
 
 use gaiku_3d::common::{mint::Vector3, Mesh};
 
-pub fn to_obj(mesh: &Mesh, position: &Vector3<f64>, name: &str) -> Object {
+pub fn to_obj(mesh: &Mesh, position: &Vector3<f32>, name: &str) -> Object {
     let mut vertices = vec![];
     let mut indices = vec![];
 
@@ -31,9 +31,9 @@ pub fn to_obj(mesh: &Mesh, position: &Vector3<f64>, name: &str) -> Object {
                 .into_iter()
                 .map(|(x, y, z)| Shape {
                     primitive: Primitive::Triangle(
-                        (x, None, None),
-                        (y, None, None),
-                        (z, None, None),
+                        (x as usize, None, None),
+                        (y as usize, None, None),
+                        (z as usize, None, None),
                     ),
                     groups: vec![],
                     smoothing_groups: vec![],
@@ -43,7 +43,7 @@ pub fn to_obj(mesh: &Mesh, position: &Vector3<f64>, name: &str) -> Object {
     }
 }
 
-pub fn export(data: Vec<(Mesh, &Vector3<f64>)>, name: &str) {
+pub fn export(data: Vec<(Mesh, &Vector3<f32>)>, name: &str) {
     let mut objects = vec![];
 
     for (index, (mesh, position)) in data.iter().enumerate() {
